resource creation end points and postman installation
