+++ package.json
@@ -4,13 +4,19 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "start": "node src/index.js",
+    "dev": "nodemon src/index.js"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "express": "^4.16.4",
     "mongodb": "^3.1.10",
-    "mongoose": "^5.3.16"
+    "mongoose": "^5.3.16",
+    "validator": "^13.9.0"
+  },
+  "devDependencies": {
+    "nodemon": "^1.19.4"
   }
 }
